+++ src/main/resources/templates/user/list.html
@@ -18,7 +18,7 @@
                 <td><a href="/users/{{userId}}">{{userId}}</a></td>
                 <td>{{name}}</td>
                 <td>{{email}}</td>
-                <td><a href="#" class="btn btn-success" role="button">수정</a></td>
+                <td><a href="/users/modify/{{userId}}" class="btn btn-success" role="button">수정</a></td>
             </tr>
             {{/users}}
             </tbody>
