+++ src/main/resources/templates/user/modify-form.html
@@ -1,0 +1,31 @@
+{{#partial "content"}}
+<div class="col-md-6 col-md-offset-3">
+    <div class="panel panel-default content-main">
+        <form action="/users/modify" method="post" name="question">
+            <div class="form-group">
+                <label for="userId">사용자 아이디</label>
+                <input class="form-control" id="userId" readonly name="userId" placeholder="User ID" value="{{user.userId}}">
+            </div>
+            <div class="form-group">
+                <label for="password">이전 비밀번호</label>
+                <input type="password" class="form-control" id="password" name="prevPassword" placeholder="Password">
+            </div>
+            <div class="form-group">
+                <label for="password">새로운 비밀번호</label>
+                <input type="password" class="form-control" id="password" name="newPassword" placeholder="Password">
+            </div>
+            <div class="form-group">
+                <label for="name">이름</label>
+                <input class="form-control" id="name" name="name" placeholder="Name" value="{{user.name}}">
+            </div>
+            <div class="form-group">
+                <label for="email">이메일</label>
+                <input type="email" class="form-control" id="email" name="email" placeholder="Email" value="{{user.email}}">
+            </div>
+            <button type="submit" class="btn btn-success clearfix pull-right">수정하기</button>
+            <div class="clearfix" />
+        </form>
+    </div>
+</div>
+{{/partial}}
+{{> base}}
